+++ CHANGELOG.md
@@ -1,20 +1,35 @@
-## 0.11.0 
+# Changelog
+
+## 0.12.0
+
+* **Breaking change:** `isDisplayed` is replaced by `hasPreviouslyCompleted`.
+* **Fix:** the `onComplete` returned value will no longer be ignored, and the next overlay won't show if this function returns `false`.
+* `FeatureDiscovery.completeCurrentStep` will force complete the current step and not execute the `onComplete` function.
+* Deprecated methods in `FeatureDiscovery` have been removed.
+* New parameter `recordStepsInSharedPreferences` in `FeatureDiscovery` to prevent step completions to be recorded in Shared Preferences.
+* New parameter `sharedPreferencesPrefix` in `FeatureDiscovery`.
+
+## 0.11.0
+
 * Added `FeatureDiscovery.backgroundOpacity` in order to customize overlay background opacity when it is displayed.
   `backgroundOpacity` is optional. If ignored, the default value is 0.96.
 
-## 0.10.0 
+## 0.10.0
+
 * Added `FeatureDiscovery.isDisplayed` method to check status of feature
-* Added Feature/animation durations customization 
+* Added Feature/animation durations customization
 * Added `openDuration` flag that controls open animation duration.
 * Added `pulseDuration` flag that controls tap target pulse animation duration.
 * Added `completeDuration` flag that controls complete animation duration.
 * Added `dismissDuration` flag that controls dismiss animation duration.
 
 ## 0.9.0
+
 * Added `barrierDismissible` flag which decides whether the overlay should dismiss
   on touching outside or not.
 
 ## 0.8.0
+
 * **Breaking change:**  `FeatureDiscovery.completeCurrentStep` is async know.
 * Added preferences for each feature to show is understood by user or not
 * Library will not show understood feature